Change type of flags to uint32_t for slurmdb_assoc_rec_t

+++ slurm/slurmdb.h
@@ -486,7 +486,7 @@ typedef struct slurmdb_assoc_rec {
 
 	uint32_t def_qos_id;       /* Which QOS id is this
 				    * associations default */
-	uint16_t flags;            /* various flags see ASSOC_FLAG_* */
+	uint32_t flags;            /* various flags see ASSOC_FLAG_* */
 	uint32_t grp_jobs;	   /* max number of jobs the
 				    * underlying group of associations can run
 				    * at one time */
